refactor: organize entity files into entities/ folder
