+++ backend/app/models/entities/tile.py
@@ -1,0 +1,6 @@
+from app.models.entities.entity import Entity
+
+
+class Tile(Entity):
+  group: int
+  tile_id: int
